New filter: protect - Write-protect parts of a plugin
It's worth providing an easy way to make devices partially read-only.
There are two main reasons for this: confining clients so they can
only write to one partition, but not to the partition table.  And
preventing clients from overwriting the start of a raw format file to
add a phony header (eg. a qcow2 header) which leads to exploits.

qemu offers something like this for "probed" raw devices, but it works
slightly differently: qemu checks all writes to the first sector
against the list of known file format signatures and prevents writes
that appear to create a different format.  This has some advantages
over what we're doing here.