file: Fix implementation of cache=none for writes
When testing virt-v2v we found that cache=none had very pessimal
performance in its current implementation when writing.  See:

  https://github.com/libguestfs/virt-v2v/commit/ac59d3b2310511b1537d408b675b19ec9a5d384e

However we know of a much better implementation - the one in nbdcopy.
This commit copies that implementation (for writes only).

A simple test is to do:

  $ ./nbdkit file out.img cache=none --run 'nbdcopy fedora-33.img $uri'

and then check the cache usage of the output file, which should be
around 0% (using https://github.com/Feh/nocache):

  $ cachestats out.img
  pages in cache: 409/1572864 (0.0%)  [filesize=6291456.0K, pagesize=4K]

For modular virt-v2v doing a local disk to local disk conversion:

 - before this change, without cache=none
   virt-v2v took 93.7 seconds, 19.1% pages cached in output file

 - before this change, enabling cache=none
   virt-v2v took 125.4 seconds, 0.0% pages cached in output file
                 ^^^ this is the bad case which caused the investigation

 - after this change, without cache=none
   virt-v2v took 93.2 seconds, 19.1% pages cached in output file

 - after this change, enabling cache=none
   virt-v2v took 97.9 seconds, 0.1% pages cached in output file

I tried to adjust NR_WINDOWS to find an optimum.  Increasing it made
no difference in performance but predictably caused a slight increase
in cached pages.  Reducing it slowed performance slightly.  So I
conclude that 8 is about right, but it probably depends on the
hardware.