tests: Use mke2fs -d to create ext4 test image
We used guestfish for this, but that was very slow (albeit working) on
armv7.  Since this is a plain ext4 filesystem without wrapper, we can
use mke2fs -d.  That was already a dependency so this doesn't change
anything.

After this change, configure tests for GUESTFISH and HAVE_GUESTFISH
are no longer used so we can remove them.