tests: Add configure --disable-libguestfs-tests flag
This can be used to disable tests which need libguestfs*.  We were
already doing that in a hackish way in the Fedora build on some
architectures.  This makes it more supportable.

Note that you can use

  ./configure --enable-libguestfs --disable-libguestfs-tests

to enable the bindings but disable the tests.

The difference between without and with the new flag on an otherwise
fully configured Fedora machine:

  # TOTAL: 286
  # PASS:  273
  # SKIP:  13

  # TOTAL: 263
  # PASS:  251
  # SKIP:  12

* except for those which directly test for requirements using
expressions like:

  requires guestfish --version

(cherry picked from commit c09ae98ff3b4b786565de4aa173274531a753d30)