cow: Include assert.h and call abort instead of assert (false)
"assert (false)" is equivalent to abort.  Apparently MinGW headers
don't define assert as a noreturn function resulting in this warning a
bit later on in the same function:

  cow.c: In function 'cow_cache':
  cow.c:610:9: error: 'mode' may be used uninitialized in this function [-Werror=maybe-uninitialized]
    610 |     r = blk_cache (next, blknum, block, mode, err);
        |         ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~

Use abort () instead of assert (false).

Also we used assert in a few other places in the file, but didn't
include <assert.h> so let's do that.