ocaml: Convert offset values correctly in several callbacks
The callbacks trim, zero, extents and cache pass an int64 offset to
the OCaml code.  However we were using caml_copy_int32 to create this
value which is obviously incorrect.  (This would work on little endian
machines for offsets < 4G, unfortunately, which is why we didn't spot
this.  Also it is not covered by any tests.)

This mistake goes right back to the original commit which added the
plugin, commit e3abac4e55 ("ocaml: New plugin lets you write plugins
as natively compiled OCaml programs.")

Fixes: commit e3abac4e5519b0d51c9a34a41af97f63920f3f32