ocaml: Use int64 as the OCaml type for counts
We previously mapped the C type uint32_t count to int32.  However this
will cause a negative number to be passed for counts >= 2G.  This is
theoretically not wrong, but could obviously cause problems for
plugins which are not prepared to deal with a negative count (by
converting it to int64 and adding 0x1_0000_0000).

As well as future-proofing the plugin, it's easier for us to just pass
counts as int64.

Note I did not change pread - see next commit.