ocaml: Pass count parameter of pread as OCaml int
The count parameter represents the length of the buffer that must be
created and returned by this function.  So OCaml int is the natural
type for this, since the size of OCaml buffers and arrays on all
platforms are represented by an int (roughly the equivalent of size_t
in C).

Note that signedness and/or overflow of the previous int32 type is not
the concern here (unlike the previous commit).  A memory buffer cannot
be larger than is represented by an OCaml int.