server/sockets: get rid of AI_ADDRCONFIG
The AI_ADDRCONFIG hint of getaddrinfo() is supposed to restrict the name
resolution to such address families (IPv4 vs. IPv6) for which the
resolving host has publicly routable addresses assigned.

The main problem with AI_ADDRCONFIG can be shown with the following
command line:

$ nbdkit -f -p 32776 -P pidfile -i ::1 --exit-with-parent null

On a host where ::1 is the only IPv6 address assigned (namely to the
loopback interface), the command fails with

> nbdkit: getaddrinfo: ::1: 32776: Address family for hostname not
> supported

due to the "publicly routable" requirement.

Remove AI_ADDRCONFIG from the getaddrinfo() hints, and as a replacement,
introduce the "-4" and "-6" options, similarly to netcat and ssh.

(1) This makes options of the form:

    -i 127.0.0.1
    -i ::1

    work regardless of "public" IPv6 / IPv4 connectivity;

(2) options of the form

    -i localhost
    -i FQDN

    will bind both IPv4 and IPv6 addresses of the desired interface(s);

(3) omitting the option "-i" will bind both IPv4 and IPv6 wildcard
    addresses (0.0.0.0 and ::);

(4) the configurations in (2) and (3) can be restricted to IPv4 or IPv6
    addresses by adding the "-4" or "-6" option, respectively.

Importantly, this change allows the "connect-tcp6" test case of libnbd to
pass on such hosts that have no IPv6 connectivity (i.e., where the only
assigned IPv6 address is ::1, namely on the loopback interface).

Ref: https://listman.redhat.com/archives/libguestfs/2022-January/msg00110.html
Signed-off-by: Laszlo Ersek <lersek@redhat.com>
Message-Id: <20220118134833.13246-3-lersek@redhat.com>
[lersek@redhat.com: fix typo in "--exit-with-parent" (Eric)]
Acked-by: Eric Blake <eblake@redhat.com>
(cherry picked from commit 9eec2335d630ae8ef947a927c1922d725d482f4a)