tests: Use a larger buffer for mlock test
This test fails sometimes on machines with 64K page sizes — see
following commits — and by using a larger buffer size we can make that
a little bit more likely.  I increased the buffer size to 10K so that
on machines with a 4K page size we are deliberately allocating a
partial page which should be rounded up by the implementation.

Reviewed-by: Laszlo Ersek <lersek@redhat.com>