Rename top level directory bash -> bash-completion
To run nbdkit without installing we recommend that you prepend the
build directory to your path:

  $ PATH=/build/nbdkit:$PATH nbdkit -fv null
  /build/nbdkit/server/nbdkit -f -v -- /build/nbdkit/plugins/null/.libs/nbdkit-null-plugin.so
  ...

This worked fine, except if you happened to run the "bash" command it
conflicted with "bash" — the name of the bash completion subdirectory.
This especially happened with the libnbd tests (because they use
“#!/usr/bin/env bash”):

  $ PATH=/build/nbdkit:$PATH make check -C copy
  make: Entering directory '/build/libnbd/copy'
  make  check-TESTS
  make[1]: Entering directory '/build/libnbd/copy'
  make[2]: Entering directory '/build/libnbd/copy'
  make[2]: bash: Permission denied
  make[2]: bash: Permission denied
  make[2]: bash: Permission denied
  make[2]: bash: Permission denied
  make[2]: bash: Permission denied
  make[2]: bash: Permission denied
  make[2]: bash: Permission denied
  make[2]: bash: Permission denied
  make[2]: bash: Permission denied
  make[2]: bash: Permission denied
  PASS: copy-nbd-to-hexdump.sh
  make[2]: bash: Permission denied
  PASS: copy-file-to-null.sh
  ...

(Oddly this doesn't stop the tests from failing)

We can avoid this awkward error/warning by renaming the bash directory
to bash-completion.