error: Tweak filter name for easier debugging
Grepping for the string 'error:' is complicated when the error filter
reports its name in the same manner.  We can keep the filter .so and
source file names the same, but tweak the filter .name field to change
messages using nbdkit_debug() to instead produce "error-inject:", so
that "error:" is reserved for uses of nbdkit_error().

Reviewed-by: Laszlo Ersek <lersek@redhat.com>