server: Add new plugin/filter .block_size callback
This commit implements a new .block_size callback in plugins and
filters.  The callback allows plugins to pass their minimum, preferred
and maximum block size through the NBD protocol NBD_INFO_BLOCK_SIZE
option.

If the client requests NBD_INFO_BLOCK_SIZE, _and_ either the plugin or
a filter provides this information, then send a reply.  Otherwise
ignore the client request.

A logically separate change was required for nbdkit-tls-filter.  It is
included in this commit to avoid breaking bisection:

    tls-fallback: Fix filter for new .block_size callback

    This filter doesn't call the next_open function in the non-TLS
    case, and therefore it never opens the plugin.  This leaves the
    internal state of nbdkit a bit strange.  There is no plugin
    context allocated, and the last filter in the chain has a context
    c_next pointer of NULL.

    This works, provided we intercept every possible callback, check
    the non-TLS case, and prevent it from calling the next function
    (because it would dereference the NULL c_next).

    To avoid a crash in backend_block_size we must therefore provide a
    .block_size callback in this filter.