nbd: In error messages, just display nbd_get_error()
The string returned by nbd_get_error() already includes the name of
the call that failed, so we don't really need to see that information
twice.  For example:

  #include <stdio.h>
  #include <libnbd.h>

  main () {
    nbd_is_read_only (nbd_create ());
    fprintf (stderr, "%s\n", nbd_get_error ());
    return 0;
  }

prints:

  nbd_is_read_only: invalid state: START: the handle must be
  negotiating, or connected with the server, or shut down: Transport
  endpoint is not connected

which is enough information to track down an error to a particular
line in this plugin.

(cherry picked from commit 00435671b40a70ae334dcffc1796633719cab1c8)