blocksize-policy: Check the correct policy
The new nbdkit-blocksize-policy-filter contained a fatal flaw in the
way it checked the policy (when blocksize-error-policy=error).  It
checked the policy of the next layer down, not the policy set by the
filter itself.

For example:

  nbdkit memory ... \
    --filter=blocksize-policy \
                blocksize-minimum=512 blocksize-error-policy=error

did not check the imposed minimum 512 byte block size.  Instead it
checked the (non-existent) constraints of the memory plugin, resulting
in this case in non-aligned requests being allowed.

This error was found when writing a test in libnbd.

Fixes: commit 1747ef0f66cf86e2a26708e9646154f85524a182