cache, cow: Export block size constraints
Because these filters perform a read-modify-write cycle for requests
which are smaller than the block size of the filter, we can adjust or
set the preferred block size to the block size of the filter or the
preferred block size of the underlying plugin, whichever is larger.

We're careful not to set a preferred block size which is larger than
the maximum block size.

After this change:

$ ./nbdkit --filter=cow floppy /usr/share/doc/nbdkit-devel \
                               cow-block-size=128k \
                               --run 'nbdinfo $uri'
protocol: newstyle-fixed without TLS
export="":
	export-size: 1458176 (1424K)
	content: DOS/MBR boot sector; partition 1 : ID=0xc, start-CHS (0x3ff,254,63), end-CHS (0x3ff,254,63), startsector 2048, 800 sectors
	uri: nbd://localhost:10809/
	contexts:
		base:allocation
	is_rotational: false
	is_read_only: false
	can_cache: true
	can_df: true
	can_fast_zero: false
	can_flush: true
	can_fua: true
	can_multi_conn: true
	can_trim: true
	can_zero: false
	block_size_minimum: 1
	block_size_preferred: 131072      <--- note
	block_size_maximum: 4294967295

Acked-by: Laszlo Ersek <lersek@redhat.com>