blocksize: Export block size constraints
This filter is a little unusual because it allows clients to send a
wider range of request sizes than the underlying plugin allows.
Therefore we advertise the widest possible minimum and maximum block
size to clients.

We still need to pick a suitable preferred block size assuming the
plugin itself doesn't advertise one.

Acked-by: Laszlo Ersek <lersek@redhat.com>