windows: server/debug.c: Don't crash if open_memstream fails
On Windows, our replacement open_memstream was failing (for other
reasons to be fixed in a later commit).  This resulted in eventually
“fputs (NULL, stderr)” being called which causes a crash.

This isn't a great situation to be in, but as it's only for a debug
message at least avoid the crash and output a partial debug message.

This also fixes the fallback path which missed "\n".