windows: Skip some nbdkit-cow-filter tests
It's unclear exactly why, but some (but not all) Windows tests of the
cow filter fail with a strange error where we are unable to truncate
the temporary overlay file.  Document the error and skip these tests
for now.

(cherry picked from commit 88bfd25c1ac42701a02f7ec526867008016dafdd)