common/include: Change unique-name macros to use __COUNTER__
Previously the macros used __LINE__ which meant they created a unique
name specific to the line on which the macro was expanded.  This
worked to a limited degree for cases like the MIN and MAX macros, but
failed in more complex cases such as when MIN and MAX were used
nested.  See also:
https://listman.redhat.com/archives/libguestfs/2022-February/msg00357.html

Use __COUNTER__ instead of __LINE__.  NBDKIT_UNIQUE_NAME must now be
used differently.  Macros must be rewritten to pass the unique symbol
name to a hidden macro, for example:

  #define FOO() \
     ({ int NBDKIT_UNIQUE_NAME(foo) = 42; \
        NBDKIT_UNIQUE_NAME(foo) * 2; })

to:

  #define FOO FOO_1(NBDKIT_UNIQUE_NAME(foo))
  #define FOO_1(foo) \
     ({ int foo = 42; \
        foo * 2; })

Thanks: Eric Blake