blocksize: Default internal block size based on plugin
Now that we have .block_size, if the plugin reports specific block
constraints, honor those by default rather than requiring the user to
duplicate work by passing in extra filter parameters on the command
line.

The TODO in test-blocksize.sh is no longer relevant (we now advertise
block size, so modern qemu no longer does read-modify-write), but that
test is still valuable as written, so the real coverage of the new
default behavior is in a new test.

Acked-by: Richard W.M. Jones <rjones@redhat.com>