common/utils: Add extensible string, based on vector
In several places we already define a string type based on a vector of
char, ie:

  DEFINE_VECTOR_TYPE(string, char);

This change simply adds a header file defining this, and updates
existing usage.  It also adds a convenient CLEANUP_FREE_STRING macro.

I called the new header "nbdkit-string.h" which is not ideal, but we
cannot use "string.h"; and "string-vector.h" is confusing because it
sounds like an argv-style array of strings.