ci: Temporarily remove check-valgrind, dist tests
check-valgrind is very specific to versions of packages installed, so
it's really only something a maintainer can run.

maintainer-check-extra-dist is a nice to have but requires invasive
changes to the EXTRA_DIST lines throughout the build system.  Martin
Kletzander has a proposed patch for this in his CI branch but I think
another approach is better, so for now disable this test.

distcheck is not expected to work at this time and will require a
great deal of work to fix.