podwrapper: Don't add the =encoding header for RHEL 6-era POD
For unclear reasons, nbdkit failed to compile any man pages on RHEL 6
with this error:

  Cannot decode string with wide characters at /usr/lib64/perl5/Encode.pm line 174.

I found by experimentation that removing the =encoding header fixed
things.  The man pages, even UTF-8 sequences, still seem to be
compiled correctly.

I based the test on the version of Perl (rather than the version of
Pod::Simple).  This is to simplify things and also because the error
comes from a built-in Perl module.