tests: Skip vsock tests unless the vsock_loopback module is loaded
It's unfortunate that vsock in Linux can be broken if some but not all
vsock modules needed are loaded.  As we are testing loopback
(VMADDR_CID_LOCAL), require that vsock_loopback is loaded.

Linux is able to autoload the correct module(s) if vsock.ko is _not_
loaded.  We could in theory enhance the test to check this, but that
leaves us with uncertainty about whether the test could work (eg. Is
vsock available as a module in this kernel?  Is this even Linux?)

Reported-by: Ming Xie
Thanks: Stefano Garzarella, Laszlo Ersek
Fixes: https://bugzilla.redhat.com/show_bug.cgi?id=2069558
(cherry picked from commit ac09992395ef0b28565a293f2c0a2855d5684d3f)