curl: scripts: Multiple fixes
Refactor duplicate code on the error path when an external cookie or
header script fails.  At the same time fix a leak of FILE* on this path.

While fixing the above, I discovered that we didn't return an error
from the do_scripts() function if the external script failed.  This is
because we tested the return value of pclose incorrectly.

Add a test to ensure this is fixed now and does not regress.

Fixes: commit dcba0f0b62ed772bf4dd4849043d8f7bec0fd127