data: Check failure from some <(...) scripts
Fix the error check in pclose.  This function returns -1 if there was
a problem running the command, or else the exit status of the command.
We only checked for -1 before, not the external command failing.

However we cannot do this in the case where we are truncating the
output of the script.  See the comment added to the code for details.