server, tests: Prefix both uses of perror ("accept")
In order to track down a very rare error in the tests of the curl
plugin header-script and cookie-script feature, I wanted to find out
if the error:

    accept: Bad file descriptor

comes from the nbdkit server or another part of the test.

To try to track that down, in the server I changed perror ("accept")
to nbdkit_error ("accept: %m").  nbdkit_error prefixes messages with
"nbdkit: error:" so it would be clear if the error was printed by the
server.  Incidental to this change I also updated a similar error path
in the same function for consistency.

We also used perror ("accept") in the minimal web server used by curl
tests.  I prefixed that error message with "web server:".

As it turns out the failure came from the web server.

General comment: We're very inconsistent in our use of perror within
the server.  This should be improved in future, but this commit
doesn't do that.

(cherry picked from commit 6320c53eca9bcfddf975a35b60be07745b63dca3)