tests: web-server: Avoid rare race in shutdown
Very rarely the curl tests could fail with:

  web server: accept: Bad file descriptor

This happened because of a race during shutdown: listen_sock was
closed in the atexit function, while the web server was blocked on
accept(2) in another thread at the same time.

This is a tricky race to fix nicely.  If you try to acquire a lock
before using listen_sock, then inevitably you get a deadlock when
cleaning up because the accept(2) path holds the lock while blocking,
preventing the cleanup() (which also wants to acquire the lock) from
progressing.  As this is only a test harness, I fixed it in a not very
nice but expedient way instead.

(cherry picked from commit f13e32b0b6fcdbeb333399c322a0ca4d6847f121)