tests: ssh: Don't run unused sshd
Commit dcc96d324f ("ssh: Add a test.") from 2019 added a test of
nbdkit-ssh-plugin.  It was supposed to start a local sshd in a
confined configuration on a randomly allocated port, and that part did
work fine.  However the test itself didn't ever connect to this
server!  The script even parsed out the random port into a local
variable which is then never used.  Instead it ran the test against
the ordinary local sshd.

It's not a problem to test with local sshd because we're just
accessing a simple local file in the test directory, and we have the
requirements set up so that it won't block if there isn't a local
server or it needs a password.

So remove all the useless machinery around this and use the normal sshd.

Fixes: commit dcc96d324ff7c8add9da1bbb1412586fe6f4b505
(cherry picked from commit c4d85791184bb316af4e279e9f937cc9eb5f9629)