readahead: Rewrite this filter so it prefetches using .cache
The previous readahead filter did not work well and we have stopped
using it in virt-v2v.  However the concept is a good one if we can
make it work.  This commit completely rethinks the filter.

Now, in parallel with the ordinary pread command, we issue a prefetch
(ie. .cache) to the underlying plugin for the data immediately
following the pread.  For example a simple sequence of operations:

  t=1  pread (offset=0, count=65536)
  t=2  pread (offset=65536, count=65536)
  t=3  pread (offset=131072, count=65536)

would become:

  t=1  pread (offset=0, count=65536)     <--\ issued
       cache (offset=65536, count=65536) <--/ in parallel
  t=2  pread (offset=65536, count=65536)
       cache (offset=131072, count=65536)
  t=3  pread (offset=131072, count=65536)
       cache (offset=196608, count=65536)

This requires that the underlying filter(s) and plugin chain can
actually do something with the .cache request.  If this is not the
case then the filter does nothing (but it will print a warning).  For
plugins which don't have native support for prefetching, it is
sufficient to insert nbdkit-cache-filter after this filter.
(nbdkit-cow-filter can also be used with cow-on-cache=true, but that
is more useful for advanced users who are already using the cow
filter).

The implementation creates a background thread per connection to issue
the parallel .cache requests.  This is safer than the alternative (one
background thread in total) since we don't have to deal with the
problem of cache requests being issued with the wrong export name or
stale next pointer.  The background thread is controlled by a queue of
commands, with the only possible commands being "cache" or "quit".

Because the background thread issues parallel requests on the same
connection, the underlying plugin must support the parallel thread
model, otherwise we would be violating the plugin's thread model.  It
may be possible in future to open a new connection from the background
thread (but with the same exportname), which would lift this
restriction to at least serialize_requests.  Because of the current
limitation, nbdkit-curl-plugin cannot use prefetch.

(cherry picked from commit 2ff548d66ad3eae87868402ec5b3319edd12090f)