readahead: Fix test
The previous test turned out to be pretty bad at testing the new
filter.  A specific problem is that the filter starts a background
thread which issues .cache requests, while on the main connection
.pread requests are being passed through.  The test used
--filter=readahead --filter=cache with the cache filter only caching
on .cache requests (since cache-on-read defaults to false), so only
caching requests made by the background thread.

                       main thread
  client ---- .pread ----- delay-filter -------> plugin
             \
              \     background thread
               .cache --- cache-filter

Under very high load, the background thread could be starved.  This
means no requests were being cached at all, and all requests were
passing through the delay filter.  It would appear that readahead was
failing (which it was, in a way).

It's not very easy to fix this since readahead is best-effort, but we
can go back to using a simpler plugin that logs reads and caches and
check that they look valid.

Update: commit 2ff548d66ad3eae87868402ec5b3319edd12090f