tests: Print a warning if nbdkit does not start after 30 seconds
When running the OCaml test under valgrind, for some reason
downloading debug symbols from debuginfod takes an extremely long time
(as in 10+ minutes).  This caused the test to fail in a rather obscure
way which was not immediately easy to diagnose.  We wait for nbdkit to
create a PID file, but if it hasn't been created within 30 seconds
then we continue anyway.  This causes a later failure (in this case
qemu being unable to connect to the NBD socket).

While this commit doesn't fix any of the underlying problems, we can
at least add a useful error message to the log file when we exceed the
timeout without the PID file having been created.

For the root cause of the valgrind problem, see commit 8c94611ca3
("wrapper: Unset DEBUGINFOD_URLS")

(cherry picked from commit 1f5dc0679e570ca2fa149ceaa52922d70462b62c)