Add nbdkit.parse_size() Python function.
This enables Python plugins to parse sizes the same way as C plugins.

I'm not sure about the best way to test this - input is appreciated.

I'm not too happy with the way this code is tested. It workes, but putting the tests into
test-python-plugin.py feels misplaced: this file is intended to support the unit tests in
test_python.py, not run its own unit tests.