nbd: Hide some state machine debugging behind a debug flag
When running virt-p2v which uses this plugin, the log file is filled
with messages about state machine transitions and so on.  In a log
file that was shared with me, of 135,023 lines in total, 94,653 (70%) were:

  nbdkit: debug: polling, dir=1

and 18,676 (14%) were:

  nbdkit: debug: cookie X completed state machine, status 0

This commit changes the logging so these state machine transitions are
only printed when you use the debug flag “-D nbd.verbose=1”.  I didn't
document this flag because it's likely only of use to developers who
are reading the code already.

Some debug messages along error paths which are (a) generally useful
and (b) did not appear in the log file, I left alone.

Fixes: https://bugzilla.redhat.com/show_bug.cgi?id=2083498
Reported-by: Ming Xie
Reviewed-by: Laszlo Ersek <lersek@redhat.com>
Reviewed-by: Eric Blake <eblake@redhat.com>