configure: Fix incorrect comment about supported VDDK version
The current minimum version is 6.5.  We document this in one place --
the nbdkit-vddk-plugin(1) manual page.

At the time of writing VDDK is not available on aarch64 (although an
experimental version of ESXi is).  We expect it will be available in
future in which case we'll need to revisit this test.