scripts: Split VDDK open script so we can open ESXi or vCenter
Move the existing, simpler script to vddk-esxi-open.sh{,.in}.  This
script does not change.

Add a new script, vddk-vcenter-open.sh{,.in} which is a modified
version of the script for doing the more complicated task of
connecting to vCenter.

I also added more debugging messages in the scripts so it's easier to
see what they are doing.