vddk: advise user on obscure thumbprint mismatch error condition
If the thumbprint parameter is wrong, it's only reported in
VixDiskLib_Open(), and then with the non-descript VIX_E_FAIL error code.
If the user typed or cut-and-pasted the thumbprint incorrectly, said
"Unknown error" message is not helpful for fixing the nbkit command line.
Hint at the thumbprint as the potential culprit.

Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=1905772
Signed-off-by: Laszlo Ersek <lersek@redhat.com>
Message-Id: <20220517080228.8713-1-lersek@redhat.com>
Acked-by: Richard W.M. Jones <rjones@redhat.com>