scan: Fix bound so we don't try to prefetch beyond end of disk
An off-by-one error in the bound could cause the filter to try to
prefetch beyond the end of the underlying plugin.  This would cause
nbdkit to crash with this assertion failure:

nbdkit: backend.c:782: backend_cache: Assertion `backend_valid_range (c, offset, count)' failed.

The sequence of events was:

 - scan filter background thread started

 - client reads to the end of the disk

 - background thread skips ahead to end of disk (offset == size)

 - background thread tries to prefetch from this point

In the final step the calculations caused to the background thread to
prefetch a scan-size block beyond the end of the plugin.

Fixes: commit 65c20a09ceacb4431986a2982f2c2e746df63fcb
(cherry picked from commit 953643429b8c57b4dd20a6c0e5b83704ae9a0e88)