ci: Try to force Rawhide to use the nodebug kernels
CI tests running on Rawhide frequently time out.  We believe this is
caused by using the debug kernels which have a lot of extra
instrumentation which makes them run very slowly under libguestfs
(particularly with TCG).

We currently have a hack in the container dockerfile which attempts to
enable the rawhide "nodebug" kernel repo.  However it's frequently the
case that those kernels are out of date with respect to Fedora
Rawhide, so the debug kernels, having higher NVRs, are always
installed.

This hack attempts to install the nodebug kernels regardless.  It's a
hack because it will be overwritten next time we run libvirt-ci, but I
just want to test out the theory.

See also:
https://lists.fedoraproject.org/archives/list/devel@lists.fedoraproject.org/thread/U4HN3P44I4JEZGM77Y7G5KUB3PMAVFXK/