blocksize: Test of aligned writes vs. RMW race
Add a test that fails without the previous change to blocksize.c.
With some carefully timed setups (using the delay filter to stall
writes reaching the plugin, and the plugin itself to stall reads
coming back from the plugin, so that we are reasonably sure of the
overall timeline), we can demonstrate the bug present in the unpatched
code, where an aligned write is lost when it lands in the wrong part
of an unaligned RMW cycle; the fixed code instead shows that the
overlapping operations were serialized.  We may need to further tweak
the test to be more reliable even under heavy load, but hopefully 2
seconds per stall (where a successful test requires 18 seconds) is
sufficient for now.

Reported-by: Nikolaus Rath <Nikolaus@rath.org>
Reviewed-by: Laszlo Ersek <lersek@redhat.com>
Reviewed-by: Richard W.M. Jones <rjones@redhat.com>
(cherry picked from commit baa63f051e22cc9c0e93fb24bd3761e93ae8c532)